Precision update
silly me for using  doubles/floats for money calculation. Changed to BigDecimal throughout project to improve precision.

+++ src/kata/products/IProduct.java
@@ -1,10 +1,12 @@
 package kata.products;
 
+import java.math.BigDecimal;
+
 import kata.products.enums.ProductType;
 
 public interface IProduct
 {
 	ProductType getProductType();
 	
-	Double getPrice();
+	BigDecimal getPrice();
 }
